Incorporate review comments: move impl class to its own pkg private class; rename the factory methods to create... instead of new...; change the accessor method names for clarity
